Make the compiled game runnable in standalone
Previously, the game could only be run as binary or exe if the resource directories were in the same absolute path. I changed it that way, so that the resources a packed up within the binary, which required the code to be changed in order to correctly pick up the resources if runin binary mode.